+++ src/drivers/Twitter.php
@@ -7,16 +7,16 @@
 
 namespace ymaker\social\share\drivers;
 
-use ymaker\social\share\base\Driver;
+use ymaker\social\share\base\DriverAbstract;
 
 /**
- * Driver for Twitter.
+ * DriverAbstract for Twitter.
  * @link https://twitter.com
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  * @since 1.0
  */
-class Twitter extends Driver
+class Twitter extends DriverAbstract
 {
     /**
      * @var bool|string
@@ -27,37 +27,40 @@ class Twitter extends Driver
     /**
      * @inheritdoc
      */
-    public function getLink()
-    {
-        $this->_link = 'http://twitter.com/share?'
-                    . 'url={url}'
-                    . '&text={description}';
-
-        if ($this->account) {
-            $this->_data['{account}'] = $this->account;
-            $this->addUrlParam('via', '{account}');
-        }
-
-        $this->_metaTags = [
-            ['name' => 'twitter:card',         'content' => 'summary_large_image'],
-            ['name' => 'twitter:title',        'content' => '{title}'],
-            ['name' => 'twitter:description',  'content' => '{description}'],
-            ['name' => 'twitter:image',        'content' => '{imageUrl}']
-        ];
-
-        return parent::getLink();
-    }
-
-    /**
-     * @inheritdoc
-     */
     protected function processShareData()
     {
         $this->url = static::encodeData($this->url);
         $this->description = static::encodeData($this->description);
 
         if (is_string($this->account)) {
-            $this->account = static::encodeData($this->account);
+            $this->appendToData('account', $this->account);
         }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function buildLink()
+    {
+        $link = 'http://twitter.com/share?url={url}&text={description}';
+
+        if ($this->account) {
+            $this->addUrlParam($link, 'via', '{account}');
+        }
+
+        return $link;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getMetaTags()
+    {
+        return [
+            ['name' => 'twitter:card',         'content' => 'summary_large_image'],
+            ['name' => 'twitter:title',        'content' => '{title}'],
+            ['name' => 'twitter:description',  'content' => '{description}'],
+            ['name' => 'twitter:image',        'content' => '{imageUrl}']
+        ];
     }
 }