+++ src/drivers/WhatsApp.php
@@ -5,12 +5,12 @@
  * @license BSD 3-Clause License
  */
 
-namespace ymaker\social\share\drivers\other\mobile;
+namespace ymaker\social\share\drivers;
 
-use ymaker\social\share\base\Driver;
+use ymaker\social\share\base\DriverAbstract;
 
 /**
- * Driver for WhatsApp messenger.
+ * DriverAbstract for WhatsApp messenger.
  * @link https://www.whatsapp.com
  *
  * WARNING: This driver works only in mobile devices
@@ -19,23 +19,21 @@ use ymaker\social\share\base\Driver;
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  * @since 1.0
  */
-class WhatsApp extends Driver
+class WhatsApp extends DriverAbstract
 {
-    /**
-     * @inheritdoc
-     */
-    public function getLink()
-    {
-        $this->_link = 'whatsapp://send?text={url}';
-
-        return parent::getLink();
-    }
-
     /**
      * @inheritdoc
      */
     protected function processShareData()
     {
         $this->url = static::encodeData($this->url);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function buildLink()
+    {
+        return 'whatsapp://send?text={url}';
     }
 }